Changed records to database directory; added directory record
